+++ resources/js/next/config/session.js
@@ -1,4 +1,5 @@
 import { getGlobalPMVariable, setGlobalPMVariables, getGlobalVariable } from "../globalVariables";
+import { initSessionSync } from "../../common/sessionSync";
 
 export default () => {
   const timeoutScript = document.head.querySelector("meta[name=\"timeout-worker\"]")?.content;
@@ -7,99 +8,46 @@ export default () => {
   const Echo = getGlobalVariable("Echo");
 
   const pushNotification = getGlobalPMVariable("pushNotification");
-  const closeSessionModal = getGlobalPMVariable("closeSessionModal");
   const alert = getGlobalPMVariable("alert");
   const user = getGlobalPMVariable("user");
+  const isProd = document.head.querySelector("meta[name=\"is-prod\"]")?.content === "true";
 
-  const isSameDevice = (e) => {
-    const localDeviceId = Vue.$cookies.get(e.device_variable);
-    const remoteDeviceId = e.device_id;
-    return localDeviceId && localDeviceId === remoteDeviceId;
-  };
-
-  if (user) {
-  // Session timeout
-    // Backend provides minutes for lifetime and seconds for warnings.
-    const AccountTimeoutLength = parseInt(eval(document.head.querySelector("meta[name=\"timeout-length\"]")?.content));
-    const warnSeconds = parseInt(document.head.querySelector("meta[name=\"timeout-warn-seconds\"]")?.content);
-    const AccountTimeoutWarnSeconds = Number.isNaN(warnSeconds) ? 0 : warnSeconds;
-    const AccountTimeoutEnabled = document.head.querySelector("meta[name=\"timeout-enabled\"]") ? parseInt(document.head.querySelector("meta[name=\"timeout-enabled\"]")?.content) : 1;
-    const AccountTimeoutWorker = new Worker(timeoutScript);
-
-    AccountTimeoutWorker.addEventListener("message", (e) => {
-      const sessionModal = getGlobalPMVariable("sessionModal");
-      if (e.data.method === "countdown") {
-        // Show countdown modal when the worker hits the warning threshold.
-        sessionModal(
-          "Session Warning",
-          "<p>Your user session is expiring. If your session expires, all of your unsaved data will be lost.</p><p>Would you like to stay connected?</p>",
-          e.data.data.time,
-          AccountTimeoutWarnSeconds,
-        );
-      }
-      if (e.data.method === "timedOut") {
-        window.location = "/logout?timeout=true";
-      }
-    });
-
-    // Start the timer for the initial page load.
-    AccountTimeoutWorker.postMessage({
-      method: "start",
-      data: {
-        timeout: AccountTimeoutLength,
-        warnSeconds: AccountTimeoutWarnSeconds,
-        enabled: AccountTimeoutEnabled,
-      },
-    });
-
-    Echo.private(`ProcessMaker.Models.User.${user.id}`)
-      .notification((token) => {
-        pushNotification(token);
-      })
-      .listen(".SessionStarted", (e) => {
-        const lifetime = parseInt(eval(e.lifetime));
-        if (isSameDevice(e)) {
-          // Reset the worker when the server renews the session.
-          AccountTimeoutWorker.postMessage({
-            method: "start",
-            data: {
-              timeout: lifetime,
-              warnSeconds: AccountTimeoutWarnSeconds,
-              enabled: AccountTimeoutEnabled,
-            },
-          });
-          if (closeSessionModal) {
-            closeSessionModal();
-          }
-        }
-      })
-      .listen(".Logout", (e) => {
-        if (isSameDevice(e) && window.location.pathname.indexOf("/logout") === -1) {
-          const localDeviceId = Vue.$cookies.get(e.device_variable);
-          const redirectLogoutinterval = setInterval(() => {
-            const newDeviceId = Vue.$cookies.get(e.device_variable);
-            if (localDeviceId !== newDeviceId) {
-              clearInterval(redirectLogoutinterval);
-              window.location.href = "/logout";
-            }
-          }, 100);
-        }
-      })
-      .listen(".SecurityLogDownloadJobCompleted", (e) => {
-        if (e.success) {
-          const { link } = e;
-          const { message } = e;
-          alert(message, "success", 0, false, false, link);
-        } else {
-          alert(e.message, "warning");
-        }
-      });
-
-    setGlobalPMVariables({
-      AccountTimeoutLength,
-      AccountTimeoutWarnSeconds,
-      AccountTimeoutEnabled,
-      AccountTimeoutWorker,
-    });
+  if (!user) {
+    return;
   }
+
+  // Backend provides minutes for lifetime and seconds for warnings.
+  const accountTimeoutLength = parseInt(eval(document.head.querySelector("meta[name=\"timeout-length\"]")?.content));
+  const warnSeconds = parseInt(document.head.querySelector("meta[name=\"timeout-warn-seconds\"]")?.content);
+  const accountTimeoutWarnSeconds = Number.isNaN(warnSeconds) ? 0 : warnSeconds;
+  const accountTimeoutEnabled = document.head.querySelector("meta[name=\"timeout-enabled\"]") ? parseInt(document.head.querySelector("meta[name=\"timeout-enabled\"]")?.content) : 1;
+
+  const sessionSyncState = initSessionSync({
+    userId: user.id,
+    isProd,
+    timeoutScript,
+    accountTimeoutLength,
+    accountTimeoutWarnSeconds,
+    accountTimeoutEnabled,
+    Vue,
+    Echo,
+    pushNotification,
+    alert,
+    getSessionModal: () => getGlobalPMVariable("sessionModal"),
+    getCloseSessionModal: () => getGlobalPMVariable("closeSessionModal"),
+    getNavbar: () => getGlobalPMVariable("navbar"),
+  });
+
+  if (!sessionSyncState) {
+    return;
+  }
+
+  setGlobalPMVariables({
+    AccountTimeoutLength: sessionSyncState.AccountTimeoutLength,
+    AccountTimeoutWarnSeconds: sessionSyncState.AccountTimeoutWarnSeconds,
+    AccountTimeoutWarnMinutes: sessionSyncState.AccountTimeoutWarnMinutes,
+    AccountTimeoutEnabled: sessionSyncState.AccountTimeoutEnabled,
+    AccountTimeoutWorker: sessionSyncState.AccountTimeoutWorker,
+    sessionSync: sessionSyncState.sessionSync,
+  });
 };
